App Code/requirements.txt
tk
Pillow
folium
geopy
openrouteservice
python-dotenv
pystray
